backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
uvicorn[standard]
SQLAlchemy==2.0.27
PyMySQL==1.1.0
cryptography==42.0.5
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart